requirements.txt
# ML & DL
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
pillow==10.0.0

# MLOps
mlflow==2.7.1
dvc[gdrive]==3.20.0

# API
fastapi==0.103.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6
pydantic==2.3.0

# Monitoring
prometheus-client==0.17.1
evidently==0.4.8

# Utils
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0
httpx==0.24.1